+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project mostly adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.0] - XXXX-XX-XX
+
+### Changed
+- Changed to use Tailwind 1.2’s new plugin definition syntax
+- Font variant utilities (caps, nums, and ligatures) are now generated from the `fontVariantCaps`, `fontVariantNumeric`, and `fontVariantLigatures` theme keys, and their names have changed to include a `caps-`, `nums-`, or `ligatures-` prefix (so `normal-caps` is now `caps-normal`, `no-ligatures` is now `ligatures-none`, etc.)
+- The `caps`, `nums`, and `ligatures` variants keys have changed to `fontVariantCaps`, `fontVariantNumeric`, and `fontVariantLigatures`
+
+### Removed
+- Removed the `caps`, `nums`, and `ligatures` options
+
 ## [2.2.0] - 2019-09-02
 
 ### Added
@@ -72,7 +82,8 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 
 Initial release
 
-[Unreleased]: https://github.com/benface/tailwindcss-typography/compare/v2.2.0...HEAD
+[Unreleased]: https://github.com/benface/tailwindcss-typography/compare/v3.0.0...HEAD
+[3.0.0]: https://github.com/benface/tailwindcss-typography/compare/v2.2.0...v3.0.0
 [2.2.0]: https://github.com/benface/tailwindcss-typography/compare/v2.1.1...v2.2.0
 [2.1.1]: https://github.com/benface/tailwindcss-typography/compare/v2.1.0...v2.1.1
 [2.1.0]: https://github.com/benface/tailwindcss-typography/compare/v2.0.0...v2.1.0